+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - **[BUGFIX]**: fix(duration): snapshot duration not set [#37](https://github.com/intergral/deep/pull/37) [@Umaaz](https://github.com/Umaaz)
 - **[BUGFIX]**: fix(attributes): snapshot attributes not set [#38](https://github.com/intergral/deep/pull/38) [@Umaaz](https://github.com/Umaaz)
 - **[BUGFIX]**: fix(resource): discovered resources overwriting each other [#41](https://github.com/intergral/deep/pull/41) [@Umaaz](https://github.com/Umaaz)
+- **[BUGFIX]**: feat(events): use events for snapshot in spans [#40](https://github.com/intergral/deep/pull/40) [@Umaaz](https://github.com/Umaaz)
 
 # 1.1.0 (06/02/2024)